@@ -13,37 +13,35 @@
   "author": "",
   "license": "ISC",
   "dependencies": {
-    "jquery": "^3.4.1",
+    "jquery": "^3.6.3",
     "normalize-scss": "^7.0.1",
-    "react": "^16.12.0",
-    "react-dom": "^16.12.0",
-    "socket.io-client": "^2.3.0"
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
+    "socket.io-client": "^4.6.1"
   },
   "devDependencies": {
-    "@babel/core": "^7.8.3",
-    "@babel/preset-env": "^7.8.3",
-    "@babel/preset-react": "^7.8.3",
-    "@hot-loader/react-dom": "^16.11.0",
-    "babel-eslint": "^10.0.3",
-    "babel-loader": "^8.0.6",
-    "css-loader": "^3.4.2",
-    "error-overlay-webpack-plugin": "^0.4.1",
-    "eslint": "^6.8.0",
-    "eslint-config-prettier": "^6.9.0",
-    "eslint-config-react": "^1.1.7",
-    "eslint-loader": "^3.0.3",
-    "eslint-plugin-react": "^7.18.0",
-    "file-loader": "^5.0.2",
-    "friendly-errors-webpack-plugin": "^1.7.0",
-    "html-webpack-plugin": "^3.2.0",
-    "node-sass": "^4.13.1",
-    "prettier": "^1.19.1",
-    "react-hot-loader": "^4.12.19",
-    "sass-loader": "^8.0.2",
-    "style-loader": "^1.1.3",
-    "url-loader": "^3.0.0",
-    "webpack": "^4.41.5",
-    "webpack-cli": "^3.3.10",
-    "webpack-dev-server": "^3.10.1"
+    "@babel/core": "^7.21.0",
+    "@babel/plugin-transform-runtime": "^7.21.0",
+    "@babel/preset-env": "^7.20.2",
+    "@babel/preset-react": "^7.18.6",
+    "@babel/runtime": "^7.21.0",
+    "babel-loader": "^9.1.2",
+    "css-loader": "^6.7.3",
+    "error-overlay-webpack-plugin": "^1.1.1",
+    "eslint": "^8.34.0",
+    "eslint-plugin-react": "^7.32.2",
+    "history": "^5.3.0",
+    "html-webpack-plugin": "^5.5.0",
+    "normalize-scss": "^7.0.1",
+    "prettier": "^2.8.4",
+    "react-refresh": "^0.14.0",
+    "react-refresh-webpack-plugin": "^0.1.0",
+    "sass": "^1.58.3",
+    "sass-loader": "^13.2.0",
+    "sharp": "^0.31.3",
+    "style-loader": "^3.3.1",
+    "webpack": "^5.75.0",
+    "webpack-cli": "^5.0.1",
+    "webpack-dev-server": "^4.11.1"
   }
 }
