@@ -1,24 +1,13 @@
 import React from 'react';
-import ReactDOM from 'react-dom';
+import { createRoot } from 'react-dom/client';
 import './style/main.scss';
-import { AppContainer } from 'react-hot-loader';
 
 import Main from './js/main';
 
 const render = Component => {
-  ReactDOM.render(
-    <AppContainer>
-      <Component />
-    </AppContainer>,
-    document.getElementById('root')
-  );
+  const container = document.getElementById('root');
+  const root = createRoot(container);
+  root.render(<Component />);
 };
 
 render(Main);
-
-if (module.hot) {
-  module.hot.accept('./js/main', () => {
-    const nextMain = require('./js/main').default;
-    render(nextMain);
-  });
-}
