@@ -1,34 +1,23 @@
 const path = require('path');
-const HtmlWebpackPlugin = require('html-webpack-plugin');
 const ErrorOverlayPlugin = require('error-overlay-webpack-plugin');
-const FriendlyErrorsWebpackPlugin = require('friendly-errors-webpack-plugin');
-const webpack = require('webpack');
+const HtmlWebpackPlugin = require('html-webpack-plugin');
 
 module.exports = {
-  entry: [
-    'react-hot-loader/patch',
-    'webpack-dev-server/client?http://localhost:3000',
-    'webpack/hot/only-dev-server',
-    './src/index.js',
-  ],
+  entry: './src/index.js',
   output: {
     path: path.resolve(__dirname, 'build'),
     filename: 'bundle.js',
-    publicPath: './',
+    publicPath: '/',
   },
   devServer: {
-    publicPath: '/',
-    contentBase: path.resolve(__dirname, 'build'),
-    stats: { colors: true },
+    static: path.resolve(__dirname, 'build'),
     host: '0.0.0.0',
     port: 3000,
     open: true,
-    inline: true,
-    hotOnly: true,
-    quiet: true,
-    overlay: {
-      warnings: true,
-    },
+    client: {
+      overlay: true,
+      progress: true,
+    }
   },
   devtool: 'cheap-module-source-map',
   module: {
@@ -36,7 +25,16 @@ module.exports = {
       {
         test: /\.(js|jsx)$/,
         exclude: /node_modules/,
-        use: ['babel-loader', 'eslint-loader'],
+        use: [
+          {
+            loader: require.resolve('babel-loader'),
+            options: {
+              // ... other options
+              // DO NOT apply the plugin in production mode!
+              plugins: [require.resolve('react-refresh/babel')],
+            },
+          },
+        ]
       },
       {
         test: /\.scss$/,
@@ -44,27 +42,18 @@ module.exports = {
       },
       {
         test: /\.(png|svg|jpg|gif)$/,
-        use: {
-          loader: 'url-loader',
-          options: {
-            limit: 25000,
-          },
-        },
+        type: 'asset'
       },
     ],
   },
   plugins: [
     new ErrorOverlayPlugin(),
-    new FriendlyErrorsWebpackPlugin(),
     new HtmlWebpackPlugin({
       template: path.resolve('./index.html'),
       filename: 'index.html',
       hash: true,
       favicon: path.resolve(__dirname, './src/assets/PAL.png'),
     }),
-    new webpack.HotModuleReplacementPlugin(),
   ],
-  resolve: {
-    alias: { 'react-dom': '@hot-loader/react-dom' },
-  },
+  mode: 'development'
 };
