@@ -2,7 +2,7 @@ const path = require('path');
 const HtmlWebpackPlugin = require('html-webpack-plugin');
 
 module.exports = {
-  entry: ['./src/index.js'],
+  entry: './src/index.js',
   output: {
     path: path.resolve(__dirname, 'build'),
     filename: 'bundle.js',
@@ -13,7 +13,7 @@ module.exports = {
       {
         test: /\.(js|jsx)$/,
         exclude: /node_modules/,
-        use: ['babel-loader', 'eslint-loader'],
+        use: ['babel-loader'],
       },
       {
         test: /\.scss$/,
@@ -21,12 +21,7 @@ module.exports = {
       },
       {
         test: /\.(png|svg|jpg|gif)$/,
-        use: {
-          loader: 'url-loader',
-          options: {
-            limit: 25000,
-          },
-        },
+        type: 'asset'
       },
     ],
   },
@@ -38,4 +33,5 @@ module.exports = {
       favicon: path.resolve(__dirname, './src/assets/PAL.png'),
     }),
   ],
+  mode: 'production'
 };
